Get normalized RNA-seq expression data

Get normalized RNA-seq expression data

rnaSeqData(
  cancerStudies,
  geneNames,
  zscore = c("all samples", "diploid samples")
)

Arguments

cancerStudies
character. Cancer study IDs. See cancerStudies() for details.

geneNames
character. HUGO gene symbols (e.g. "TP53").

zscore
character.

"all samples": mRNA expression z-Scores relative to all samples (log RNA Seq RPKM). Log-transformed mRNA z-Scores compared to the expression distribution of all samples (RNA Seq RPKM).

"diploid samples": mRNA expression z-Scores relative to diploid samples (RNA Seq RPKM). mRNA z-Scores (RNA Seq RPKM) compared to the expression distribution of each gene tumors that are diploid for this gene.

Value

DataFrame.

Details

Examples of cancer studies with different mRNA data types:

RNA-seq v2: gbm_tcga_pub2013_rna_seq_v2_mrna.

RNA-seq v1: nbl_target_2018_pub_rna_seq_mrna.

Microarray: gbm_tcga_pub_mrna.

Note

Updated 2020-10-09.

The cBioPortal Z-Score calculation method

cBioPortal currently generates two z-score profiles using two different base populations:

Distribution based on diploid samples only: The expression distribution for unaltered copies of the gene is estimated by calculating the mean and variance of the expression values for samples in which the gene is diploid (i.e. value is "0" as reported by discrete CNA data). We call this the unaltered distribution. If the gene has no diploid samples, then its normalized expression is reported as NA.

Distribution based on all samples: The expression distribution of the gene is estimated by calculating the mean and variance of all samples with expression values (excludes zero's and non-numeric values like NA, NULL or NaN). If the gene has samples whose expression values are all zeros or non-numeric, then its normalized expression is reported as NA.

Otherwise for every sample, the gene's normalized expression for both the profiles is reported as:

(r - mu)/sigma

where r is the raw expression value, and mu and sigma are the mean and standard deviation of the base population, respectively.

See also: https://github.com/cBioPortal/cbioportal/blob/master/docs/Z-Score-normalization-script.md

Examples

cancerStudies <- c("acc_tcga_pan_can_atlas_2018", "ccle_broad_2019")
geneNames <- c("MYC", "TP53")
df <- rnaSeqData(cancerStudies = cancerStudies, geneNames = geneNames)

#> → Importing RNA-seq data: `acc_tcga_pan_can_atlas_2018_rna_seq_v2_mrna_median_all_sample_Zscores`.

#> → Importing RNA-seq data: `ccle_broad_2019_rna_seq_mrna_median_all_sample_Zscores`.

print(df)

#> DataFrame with 1234 rows and 4 columns
#>                      cancerStudy                      sampleID       MYC
#>                      <character>                   <character> <numeric>
#> 1    acc_tcga_pan_can_atlas_2018               TCGA.OR.A5J1.01   -0.2656
#> 2    acc_tcga_pan_can_atlas_2018               TCGA.OR.A5J2.01    0.5668
#> 3    acc_tcga_pan_can_atlas_2018               TCGA.OR.A5J3.01   -0.4335
#> 4    acc_tcga_pan_can_atlas_2018               TCGA.OR.A5J5.01   -0.2846
#> 5    acc_tcga_pan_can_atlas_2018               TCGA.OR.A5J6.01    0.0136
#> ...                          ...                           ...       ...
#> 1230             ccle_broad_2019 YD8_UPPER_AERODIGESTIVE_TRACT   -0.4368
#> 1231             ccle_broad_2019   YH13_CENTRAL_NERVOUS_SYSTEM   -0.9147
#> 1232             ccle_broad_2019   YKG1_CENTRAL_NERVOUS_SYSTEM   -1.0933
#> 1233             ccle_broad_2019                  ZR751_BREAST    0.1006
#> 1234             ccle_broad_2019                 ZR7530_BREAST    0.1984
#>           TP53
#>      <numeric>
#> 1      -0.6076
#> 2      -3.1219
#> 3       0.4805
#> 4       0.8948
#> 5      -0.8847
#> ...        ...
#> 1230    1.1431
#> 1231   -0.3186
#> 1232    0.8817
#> 1233   -0.4317
#> 1234   -0.2574