Get normalized RNA-seq expression data

Get normalized RNA-seq expression data

rnaSeqData(
  cancerStudies,
  geneNames,
  zscore = c("all samples", "diploid samples")
)

Arguments

cancerStudies
character. Cancer study IDs. See cancerStudies() for details.

geneNames
character. HUGO gene symbols (e.g. "TP53").

zscore
character.

"all samples": mRNA expression z-Scores relative to all samples (log RNA Seq RPKM). Log-transformed mRNA z-Scores compared to the expression distribution of all samples (RNA Seq RPKM).

"diploid samples": mRNA expression z-Scores relative to diploid samples (RNA Seq RPKM). mRNA z-Scores (RNA Seq RPKM) compared to the expression distribution of each gene tumors that are diploid for this gene.

Value

DataFrame.

Details

Examples of cancer studies with different mRNA data types:

RNA-seq v2: gbm_tcga_pub2013_rna_seq_v2_mrna.

RNA-seq v1: nbl_target_2018_pub_rna_seq_mrna.

Microarray: gbm_tcga_pub_mrna.

Note

Updated 2020-10-09.

The cBioPortal Z-Score calculation method

cBioPortal currently generates two z-score profiles using two different base populations:

Distribution based on diploid samples only: The expression distribution for unaltered copies of the gene is estimated by calculating the mean and variance of the expression values for samples in which the gene is diploid (i.e. value is "0" as reported by discrete CNA data). We call this the unaltered distribution. If the gene has no diploid samples, then its normalized expression is reported as NA.

Distribution based on all samples: The expression distribution of the gene is estimated by calculating the mean and variance of all samples with expression values (excludes zero's and non-numeric values like NA, NULL or NaN). If the gene has samples whose expression values are all zeros or non-numeric, then its normalized expression is reported as NA.

Otherwise for every sample, the gene's normalized expression for both the profiles is reported as:

(r - mu)/sigma

where r is the raw expression value, and mu and sigma are the mean and standard deviation of the base population, respectively.

See also: https://github.com/cBioPortal/cbioportal/blob/master/docs/Z-Score-normalization-script.md

Examples

cancerStudies <- c("acc_tcga_pan_can_atlas_2018", "ccle_broad_2019")
geneNames <- c("MYC", "TP53")
df <- rnaSeqData(cancerStudies = cancerStudies, geneNames = geneNames)
print(df)

#> DataFrame with 1234 rows and 4 columns
#>                      cancerStudy                     sampleID       MYC
#>                      <character>                  <character> <logical>
#> 1    acc_tcga_pan_can_atlas_2018                   A101D_SKIN        NA
#> 2    acc_tcga_pan_can_atlas_2018 A1207_CENTRAL_NERVOUS_SYSTEM        NA
#> 3    acc_tcga_pan_can_atlas_2018  A172_CENTRAL_NERVOUS_SYSTEM        NA
#> 4    acc_tcga_pan_can_atlas_2018             A204_SOFT_TISSUE        NA
#> 5    acc_tcga_pan_can_atlas_2018                   A2058_SKIN        NA
#> ...                          ...                          ...       ...
#> 1230             ccle_broad_2019              TCGA.PA.A5YG.01        NA
#> 1231             ccle_broad_2019              TCGA.PK.A5H8.01        NA
#> 1232             ccle_broad_2019              TCGA.PK.A5H9.01        NA
#> 1233             ccle_broad_2019              TCGA.PK.A5HA.01        NA
#> 1234             ccle_broad_2019              TCGA.PK.A5HB.01        NA
#>           TP53
#>      <logical>
#> 1           NA
#> 2           NA
#> 3           NA
#> 4           NA
#> 5           NA
#> ...        ...
#> 1230        NA
#> 1231        NA
#> 1232        NA
#> 1233        NA
#> 1234        NA